+++ packages/vuikit/src/icon/components/icon-button.js
@@ -1,15 +1,15 @@
-import { mergeData } from 'vuikit/src/_core/utils/vue'
+import { assign } from 'vuikit/src/_core/utils/object'
 
-import core from './core'
+import core from '../core'
 import { ElIconButton } from '../elements'
 
 export default {
   name: 'VkIconButton',
   functional: true,
-  props: core.props,
+  props: assign({}, core.props, ElIconButton.props),
   render (h, { data, props }) {
     return h(ElIconButton, data, [
-      h(core, mergeData(data, { props }))
+      h(core, { attrs: data.attrs, props })
     ])
   }
 }