+++ packages/vuikit/src/icon/components/icon-link.js
@@ -1,7 +1,6 @@
 import { assign } from 'vuikit/src/_core/utils/object'
-import { mergeData } from 'vuikit/src/_core/utils/vue'
 
-import core from './core'
+import core from '../core'
 import { ElIconLink } from '../elements'
 
 export default {
@@ -9,7 +8,8 @@ export default {
   functional: true,
   props: assign({}, core.props, ElIconLink.props),
   render (h, { data, props }) {
-    const def = mergeData(data, { props })
-    return h(ElIconLink, def, [ h(core, def) ])
+    return h(ElIconLink, data, [
+      h(core, { attrs: data.attrs, props })
+    ])
   }
 }