+++ packages/vuikit/src/icon/index.js
@@ -1,4 +1,5 @@
+import core from './core'
 import * as elements from './elements'
 
-export { elements }
+export { elements, core }
 export * from './components'