+++ packages/vuikit/src/iconnav/components/iconnav--item.js
@@ -1,20 +1,16 @@
 import { assign } from 'vuikit/src/_core/utils/object'
 import { mergeData } from 'vuikit/src/_core/utils/vue'
 
+import { core } from 'vuikit/src/icon'
 import { ElIconnavItem } from '../elements'
 
 export default {
   name: 'VkIconnavItem',
   functional: true,
-  props: assign({
-    icon: {
-      type: String,
-      required: true
-    }
-  }, ElIconnavItem.props),
+  props: assign({}, core.props, ElIconnavItem.props),
   render (h, { data, props }) {
     return h(ElIconnavItem, mergeData(data, { props }), [
-      h(`vk-icons-${props.icon}`)
+      h(core, { attrs: data.attrs, props })
     ])
   }
 }