+++ packages/vuikit/src/nav/components/nav--item.js
@@ -1,19 +1,18 @@
 import { assign } from 'vuikit/src/_core/utils/object'
 import { mergeData } from 'vuikit/src/_core/utils/vue'
 
+import { core } from 'vuikit/src/icon'
 import { ElNavItem } from '../elements'
 
 export default {
   name: 'VkNavItem',
   functional: true,
-  props: assign({}, ElNavItem.props, {
-    icon: {
-      type: String
-    }
-  }),
+  props: assign({}, core.props, {
+    icon: { required: false }
+  }, ElNavItem.props),
   render (h, { data, props }) {
     return h(ElNavItem, mergeData(data, { props }), [
-      props.icon && h(`vk-icons-${props.icon}`, { slot: 'icon' })
+      props.icon && h(core, { props, slot: 'icon' })
     ])
   }
 }