+++ packages/vuikit/src/tabs/components/tabs--item.js
@@ -1,22 +1,19 @@
 import { assign } from 'vuikit/src/_core/utils/object'
 import { mergeData } from 'vuikit/src/_core/utils/vue'
 
+import { core } from 'vuikit/src/icon'
 import { TAB_ID } from '../constants'
 import { ElTabsItem } from '../elements'
 
 export default {
   name: 'VkTabsItem',
-  props: assign({}, ElTabsItem.props, {
-    icon: {
-      type: String
-    }
-  }),
+  props: assign({}, core.props, {
+    icon: { required: false }
+  }, ElTabsItem.props),
   render (h) {
     return h('div', this.$slots.default)
   },
-  tabRender (h, { data, props, children, parent }) {
-    props.icon = props.icon && h(`vk-icons-${props.icon}`)
-
+  tabRender (h, { data, props, children = [], parent }) {
     return h(ElTabsItem, mergeData(data, {
       props,
       on: {
@@ -25,6 +22,9 @@ export default {
           parent.setActiveTab(data[TAB_ID])
         }
       }
-    }), children)
+    }), [
+      props.icon && h(core, { props, slot: 'icon' }),
+      ...children
+    ])
   }
 }